% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avian_common_count}
\alias{avian_common_count}
\title{Avian common count data}
\format{A data.frame with 3 variables: \code{survey}, \code{abbrev},
\code{count}}
\usage{
avian_common_count
}
\description{
For 16 species, the number of individuals of that species observed during
each survey with zero counts removed from the data set. To include zeros,
see the examples below.
}
\examples{
library(dplyr)
library(tidyr)
with_zeros <- avian_common_count \%>\%
  right_join(avian_surveys, by="survey") \%>\%     # include all surveys
  complete(survey, abbrev, fill = list(count=0)) # fill with zero counts
}
\seealso{
\code{\link{avian_surveys}}, \code{\link{avian_abbrev}}
}
\keyword{datasets}

