% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avian_groups}
\alias{avian_groups}
\title{Avian groups}
\format{A data.frame with three variables: \code{abbrev},
\code{grassland_species}, and \code{species_of_greatest_conservation_need}}
\usage{
avian_groups
}
\description{
Whether a species is a grassland species and whether the species is of
greatest conservation need.
}
\keyword{datasets}

