% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avian_surveys}
\alias{avian_surveys}
\title{Avian surveys}
\format{A data.frame with 3 variables: \code{survey}, \code{date}, and
\code{watershed}}
\usage{
avian_surveys
}
\description{
A list of all avian surveys performed during the STRIPS project. Each survey
corresponds to a particular date-watershed combination.
}
\examples{
library(dplyr)
avian_community \%>\%
  left_join(avian_surveys, by="survey")
}
\seealso{
\code{\link{avian_common_count}}, \code{\link{avian_community}}
}
\keyword{datasets}

