% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRIPSSchulte.R
\docType{package}
\name{STRIPSSchulte}
\alias{STRIPSSchulte}
\alias{STRIPSSchulte-package}
\title{STRIPSSchulte: A package for avian counts from the lab of Lisa Schulte-Moore.}
\description{
This package contains counts on 16 relatively abundant species as well as
community abundance, richness, and diversity measurements.
}
\section{STRIPSSchulte data sets}{

\code{\link{avian_abbrev}}
\code{\link{avian_common_count}}
\code{\link{avian_community}}
\code{\link{avian_groups}}
\code{\link{avian_surveys}}
}
\seealso{
\code{\link{avian_abbrev}}, \code{\link{avian_common_count}},
\code{\link{avian_community}}, \code{\link{avian_groups}},
\code{\link{avian_surveys}}
}

