% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avian_abbrev}
\alias{avian_abbrev}
\title{Avian codes}
\format{A data.frame with two variables: \code{abbrev} and
\code{common_name}}
\usage{
avian_abbrev
}
\description{
Each species is given a 4 digit character abbreviation This data sets
provides the common name associated with each abbreviation.
}
\keyword{datasets}

