% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avian_community}
\alias{avian_community}
\title{Avian community abundance, richness, and diversity}
\format{A data.frame with 7 variables: \code{survey}, \code{year},
\code{julian}, \code{watershed}, \code{abundance}, \code{richness}, and
Simpson's inverse \code{diversity}}
\usage{
avian_community
}
\description{
For each survey, abundance, richness, and diversity measurements.
}
\keyword{datasets}

